Add environment support
Add support for environment directives. See
https://www.chordpro.org/chordpro/directives-env/
for a list of environments.

Currently the only supported environments are the ones listed in
lib/chordpro/setup.rb, though by the chordpro specification any
environment name delimited by {start_of_<name>} and {end_of_<name>}
should be allowed. This is something to be added in a later pull
request.